test: rename random generator test files and split large integration suite
